backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.30.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.4.2

# Data processing
pandas==2.1.4
numpy==1.26.4
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
lightgbm==4.2.0

# Geo / Utils
geojson==3.1.0
pyproj==3.6.1
shapely==2.0.2
